+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Linear Changelog
 
+## [Added two new actions] - 2024-05-10
+
+- Added two new toast actions: `Copy Issue ID as Link` and `Copy Issue Title as Link`
+
 ## [Fix missing icons] - 2024-05-01
 
 - Update missing Linear icons